QnA , FAQ , NOTICE 패키지 CRUD 메소드 추가
+ 그래들리 querydsl 라이브러리 추가

+++ src/main/java/com/crud/btt/cs/entity/QnAEntity.java
@@ -38,5 +38,7 @@ public class QnAEntity {
     @Column(name="QNA_PRIVATE")
     private String qna_private;
     @Column(name="QNA_REF")
-    private int qna_ref;
+    private long qna_ref;
+
+
 }
